@@ -1,26 +1,26 @@
 package fr.pinguet62.test.jsr303;
 
-import static java.util.Arrays.asList;
-
-import java.util.ArrayList;
-import java.util.List;
+import org.external.ExternalModel;
+import org.hibernate.validator.constraints.Length;
+import org.junit.jupiter.api.Test;
 
 import javax.validation.ConstraintViolationException;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.util.ArrayList;
+import java.util.List;
 
-import org.external.ExternalModel;
-import org.hibernate.validator.constraints.Length;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
-/** @see Jsr303ValidationAspect */
-public class Jsr303ValidationAspectTest {
+class Jsr303ValidationAspectTest {
 
-    public static class Service {
-        private final List<String> result;
+    static class Service {
+        final List<String> result;
 
-        /** Constructor to mock {@link #getEmail(String)}. */
+        /**
+         * Constructor to mock {@link #getEmail(String)}.
+         */
         public Service(List<String> result) {
             this.result = result;
         }
@@ -32,7 +32,7 @@ public class Jsr303ValidationAspectTest {
         }
     }
 
-    public static class ConstructorClass {
+    static class ConstructorClass {
         public ConstructorClass(@Length(min = 2, max = 5) String arg) {
         }
     }
@@ -43,60 +43,59 @@ public class Jsr303ValidationAspectTest {
      * @see Jsr303ValidationAspect#intoApplication()
      */
     @Test
-    public void test_packageFilter() {
+    void test_packageFilter() {
         new ExternalModel().requireNotNull(null);
     }
 
     @Test
-    public void test_constructor_ok() {
+    void test_constructor_ok() {
         new ConstructorClass("any");
     }
 
-    @Test(expected = ConstraintViolationException.class)
-    public void test_constructor_error_Length() {
-        new ConstructorClass("");
-    }
-
-    @Test(expected = ConstraintViolationException.class)
-    public void test_parameter_error_NotNull() {
-        Service serice = new Service(asList("foo@bar.org"));
-        serice.getEmail(null);
-    }
-
-    @Test(expected = ConstraintViolationException.class)
-    public void test_parameter_error_Length_min() {
-        Service serice = new Service(asList("foo@bar.org"));
-        serice.getEmail("");
-    }
-
-    @Test(expected = ConstraintViolationException.class)
-    public void test_parameter_error_Length_max() {
-        Service serice = new Service(asList("foo@bar.org"));
-        serice.getEmail("xxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    @Test
+    void test_constructor_error_Length() {
+        assertThrows(ConstraintViolationException.class, () -> new ConstructorClass(""));
     }
 
     @Test
-    public void test_return_ok() {
-        Service serice = new Service(asList("foo@bar.org"));
+    void test_parameter_error_NotNull() {
+        Service serice = new Service(List.of("foo@bar.org"));
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail(null));
+    }
+
+    @Test
+    void test_parameter_error_Length_min() {
+        Service serice = new Service(List.of("foo@bar.org"));
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail(""));
+    }
+
+    @Test
+    void test_parameter_error_Length_max() {
+        Service serice = new Service(List.of("foo@bar.org"));
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail("xxxxxxxxxxxxxxxxxxxxxxxxxxx"));
+    }
+
+    @Test
+    void test_return_ok() {
+        Service serice = new Service(List.of("foo@bar.org"));
         serice.getEmail("any");
     }
 
-    @Test(expected = ConstraintViolationException.class)
-    public void test_return_error_NotNull() {
+    @Test
+    void test_return_error_NotNull() {
         Service serice = new Service(null);
-        serice.getEmail("any");
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail("any"));
     }
 
-    @Test(expected = ConstraintViolationException.class)
-    public void test_return_error_Size() {
+    @Test
+    void test_return_error_Size() {
         Service serice = new Service(new ArrayList<>());
-        serice.getEmail("any");
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail("any"));
     }
 
-    @Test(expected = ConstraintViolationException.class)
-    public void test_return_generic_error_Email() {
-        Service serice = new Service(asList("invalid"));
-        serice.getEmail("any");
+    @Test
+    void test_return_generic_error_Email() {
+        Service serice = new Service(List.of("invalid"));
+        assertThrows(ConstraintViolationException.class, () -> serice.getEmail("any"));
     }
-
 }
