@@ -1,54 +1,50 @@
 package fr.pinguet62.test.springboot.routingdatasource;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
 
-public class OnPropertyGroupConditionTest {
+class OnPropertyGroupConditionTest {
 
     @Configuration
     @ConditionalOnPropertyGroup("root.sub")
-    protected static class ConditionedConfig {
+    static class ConditionedConfig {
         @Bean
-        public String foo() {
+        String foo() {
             return "bar";
         }
     }
 
-    private static final String BEAN = "foo";
+    static final String BEAN = "foo";
 
-    private final ApplicationContextRunner contextRunner = new ApplicationContextRunner().withUserConfiguration(ConditionedConfig.class);
+    final ApplicationContextRunner contextRunner = new ApplicationContextRunner().withUserConfiguration(ConditionedConfig.class);
 
     @Test
-    public void emptyProperties_false() {
+    void emptyProperties_false() {
         contextRunner.withPropertyValues(
                 // no config
         ).run(context ->
-                assertThat(context.containsBean(BEAN), is(false))
-        );
+                assertThat(context.containsBean(BEAN), is(false)));
     }
 
     @Test
-    public void withoutSubAttributes_false() {
+    void withoutSubAttributes_false() {
         contextRunner.withPropertyValues(
                 "root.sub=X"
         ).run(context ->
-                assertThat(context.containsBean(BEAN), is(false))
-        );
+                assertThat(context.containsBean(BEAN), is(false)));
     }
 
     @Test
-    public void withSubAttributes_true() {
+    void withSubAttributes_true() {
         contextRunner.withPropertyValues(
                 "root.sub.attr1=A",
                 "root.sub.attr2=B"
         ).run(context ->
-                assertThat(context.containsBean(BEAN), is(true))
-        );
+                assertThat(context.containsBean(BEAN), is(true)));
     }
-
 }
