@@ -2,24 +2,21 @@ package fr.pinguet62.test.springdatathrowifnotfound;
 
 import fr.pinguet62.test.springdatathrowifnotfound.config.NotFoundException;
 import fr.pinguet62.test.springdatathrowifnotfound.sample.SampleRepository;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
-public class SampleTest {
+class SampleTest {
 
     @Autowired
-    private SampleRepository repository;
+    SampleRepository repository;
 
     @Test
-    public void test_findByIdOrThrow() {
+    void test_findByIdOrThrow() {
         NotFoundException error = null;
         try {
             repository.findByIdOrThrow(42);
@@ -30,12 +27,12 @@ public class SampleTest {
     }
 
     @Test
-    public void test_generatedMethod_notAnnotated() {
+    void test_generatedMethod_notAnnotated() {
         assertNull(repository.findByIdOrName(42, "unknown"));
     }
 
     @Test
-    public void test_generatedMethod_annotated() {
+    void test_generatedMethod_annotated() {
         NotFoundException error = null;
         try {
             repository.findByName("unknown");
@@ -44,5 +41,4 @@ public class SampleTest {
         }
         assertNotNull(error);
     }
-
 }
