@@ -1,7 +1,7 @@
 package fr.pinguet62.test.springsecurityjwt.web;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -15,9 +15,9 @@ import javax.servlet.FilterChain;
 
 import static fr.pinguet62.test.springsecurityjwt.web.JwtAuthenticationFilter.HEADER_KEY;
 import static fr.pinguet62.test.springsecurityjwt.web.JwtAuthenticationFilter.TOKEN_PREFIX;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
-import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
@@ -25,22 +25,22 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class JwtAuthenticationFilterTest {
+class JwtAuthenticationFilterTest {
 
-    private AuthenticationManager authenticationManager;
-    private AuthenticationEntryPoint authenticationEntryPoint;
-    private JwtAuthenticationFilter filter;
-    private MockHttpServletRequest request;
-    private MockHttpServletResponse response;
-    private FilterChain filterChain;
+    AuthenticationManager authenticationManager;
+    AuthenticationEntryPoint authenticationEntryPoint;
+    JwtAuthenticationFilter filter;
+    MockHttpServletRequest request;
+    MockHttpServletResponse response;
+    FilterChain filterChain;
 
-    @Before
-    public void clearSecurityContext() {
+    @BeforeEach
+    void clearSecurityContext() {
         SecurityContextHolder.clearContext();
     }
 
-    @Before
-    public void initMocks() {
+    @BeforeEach
+    void initMocks() {
         request = new MockHttpServletRequest();
         response = new MockHttpServletResponse();
         filterChain = mock(FilterChain.class);
@@ -50,7 +50,7 @@ public class JwtAuthenticationFilterTest {
     }
 
     @Test
-    public void noHeader_continue() throws Exception {
+    void noHeader_continue() throws Exception {
         filter.doFilterInternal(request, response, filterChain);
 
         verify(filterChain).doFilter(request, response);
@@ -58,7 +58,7 @@ public class JwtAuthenticationFilterTest {
     }
 
     @Test
-    public void malFormattedHeader_continue() throws Exception {
+    void malFormattedHeader_continue() throws Exception {
         request.addHeader(HEADER_KEY, "not a JWT token");
 
         filter.doFilterInternal(request, response, filterChain);
@@ -68,7 +68,7 @@ public class JwtAuthenticationFilterTest {
     }
 
     @Test
-    public void malFormattedTokenAndFailureNotIgnored_callsAuthenticationEntryPointAndStop() throws Exception {
+    void malFormattedTokenAndFailureNotIgnored_callsAuthenticationEntryPointAndStop() throws Exception {
         request.addHeader(HEADER_KEY, TOKEN_PREFIX + "bad token");
 
         filter.doFilterInternal(request, response, filterChain);
@@ -78,7 +78,7 @@ public class JwtAuthenticationFilterTest {
     }
 
     @Test
-    public void authenticationFailsAndFailureNotIgnored_callsAuthenticationEntryPointAndStop() throws Exception {
+    void authenticationFailsAndFailureNotIgnored_callsAuthenticationEntryPointAndStop() throws Exception {
         request.addHeader(HEADER_KEY, TOKEN_PREFIX + "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c");
         when(authenticationManager.authenticate(any(Authentication.class))).thenThrow(new AuthenticationServiceException("..."));
 
@@ -89,7 +89,7 @@ public class JwtAuthenticationFilterTest {
     }
 
     @Test
-    public void authenticationSuccess_continue() throws Exception {
+    void authenticationSuccess_continue() throws Exception {
         request.addHeader(HEADER_KEY, TOKEN_PREFIX + "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c");
         Authentication authentication = mock(Authentication.class);
         when(authenticationManager.authenticate(any(Authentication.class))).thenReturn(authentication);
@@ -99,5 +99,4 @@ public class JwtAuthenticationFilterTest {
         verify(filterChain).doFilter(request, response);
         assertThat(SecurityContextHolder.getContext().getAuthentication(), is(authentication));
     }
-
 }
