@@ -1,38 +1,34 @@
 package fr.pinguet62.test.springstatemachine.errorhandling.custom;
 
-import static java.util.EnumSet.allOf;
-import static org.junit.Assert.fail;
-import static org.springframework.statemachine.config.StateMachineBuilder.builder;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import fr.pinguet62.test.springstatemachine.errorhandling.InternalStateMachineException;
+import fr.pinguet62.test.springstatemachine.errorhandling.MyEvent;
+import fr.pinguet62.test.springstatemachine.errorhandling.MyState;
+import fr.pinguet62.test.springstatemachine.errorhandling.TestConfig;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.statemachine.StateMachine;
 import org.springframework.statemachine.action.Action;
 import org.springframework.statemachine.config.StateMachineBuilder.Builder;
-import org.springframework.test.context.junit4.SpringRunner;
 
-import fr.pinguet62.test.springstatemachine.errorhandling.InternalStateMachineException;
-import fr.pinguet62.test.springstatemachine.errorhandling.MyEvent;
-import fr.pinguet62.test.springstatemachine.errorhandling.MyState;
-import fr.pinguet62.test.springstatemachine.errorhandling.TestConfig;
+import static java.util.EnumSet.allOf;
+import static org.junit.jupiter.api.Assertions.fail;
+import static org.springframework.statemachine.config.StateMachineBuilder.builder;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest(classes = TestConfig.class)
-public class ActivityFailTest {
+class ActivityFailTest {
 
     @Autowired
-    private BeanFactory beanFactory;
+    BeanFactory beanFactory;
 
     @Autowired
-    private Action<MyState, MyEvent> fallingAction;
+    Action<MyState, MyEvent> fallingAction;
 
     @Autowired
-    private MyService service;
+    MyService service;
 
-    private StateMachine<MyState, MyEvent> build() throws Exception {
+    StateMachine<MyState, MyEvent> build() throws Exception {
         Builder<MyState, MyEvent> builder = builder();
         // @formatter:off
         builder.configureConfiguration()
@@ -53,7 +49,7 @@ public class ActivityFailTest {
     }
 
     @Test
-    public void test() throws Exception {
+    void test() throws Exception {
         StateMachine<MyState, MyEvent> stateMachine = build();
 
         try {
@@ -63,5 +59,4 @@ public class ActivityFailTest {
             // ok
         }
     }
-
 }
