@@ -1,7 +1,7 @@
-package fr.pinguet62.test.pattern.specification.rule;
+package fr.pinguet62.test.pattern.specification.api.rule;
 
 import fr.pinguet62.test.pattern.specification.Specification;
-import fr.pinguet62.test.pattern.specification.model.Product;
+import fr.pinguet62.test.pattern.specification.api.model.Product;
 
 public class IsBlack implements Specification<Product> {
 
@@ -9,5 +9,4 @@ public class IsBlack implements Specification<Product> {
     public boolean isSatisfiedBy(Product product) {
         return product.getColor().equals("black");
     }
-
 }
