@@ -1,9 +1,8 @@
 package fr.pinguet62.test.springbootadmin;
 
+import de.codecentric.boot.admin.server.config.EnableAdminServer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import de.codecentric.boot.admin.server.config.EnableAdminServer;
 
 @SpringBootApplication
 @EnableAdminServer
@@ -12,5 +11,4 @@ public class SpringBootAdminServerApplication {
     public static void main(String[] args) {
         SpringApplication.run(SpringBootAdminServerApplication.class, args);
     }
-
 }
