@@ -1,7 +1,7 @@
 package fr.pinguet62.test.springdatathrowifnotfound.sample;
 
-import fr.pinguet62.test.springdatathrowifnotfound.config.ThrowIfNotFoundJpaRepository;
 import fr.pinguet62.test.springdatathrowifnotfound.config.ThrowIfNotFound;
+import fr.pinguet62.test.springdatathrowifnotfound.config.ThrowIfNotFoundJpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -11,5 +11,4 @@ public interface SampleRepository extends ThrowIfNotFoundJpaRepository<SampleEnt
     SampleEntity findByName(String name);
 
     SampleEntity findByIdOrName(Integer id, String name);
-
 }
