@@ -3,4 +3,5 @@ package fr.pinguet62.test.messagesourcedatabase;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DatabaseMessageSourceConfig {}
+public class DatabaseMessageSourceConfig {
+}
