@@ -1,17 +1,16 @@
 package fr.pinguet62.test.messagesourcedatabase;
 
-import java.util.Locale;
-import java.util.Set;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.Locale;
+import java.util.Set;
 
 @Repository
 public interface MessageRepository extends CrudRepository<Message, String>, MessageRepositoryCustom {
 
     Message findValueByCodeAndLocale(String code, String locale);
-
 }
 
 interface MessageRepositoryCustom {
@@ -19,15 +18,14 @@ interface MessageRepositoryCustom {
     /**
      * @param locale Ordered {@link Locale}.
      * @return The {@link Message#getValue()}<br>
-     *         {@code null} if not found.
+     * {@code null} if not found.
      */
     String findValueByCodeAndLocaleOrDefault(String code, Set<String> locales);
-
 }
 
 class MessageRepositoryImpl implements MessageRepositoryCustom {
 
-    @Autowired
+    @Autowired // workaround: circular reference
     private MessageRepository messageRepository;
 
     @Override
@@ -39,5 +37,4 @@ class MessageRepositoryImpl implements MessageRepositoryCustom {
         }
         return null;
     }
-
 }
