@@ -1,5 +1,7 @@
 package fr.pinguet62.test.springreactiveglobalcontext.beanscoperequest;
 
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
@@ -10,14 +12,7 @@ import static org.springframework.web.context.WebApplicationContext.SCOPE_REQUES
 @Scope(value = SCOPE_REQUEST, proxyMode = TARGET_CLASS)
 public class BeanScopeRequestStorage {
 
+    @Getter
+    @Setter
     private String value;
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
 }
