@@ -1,7 +1,12 @@
 package fr.pinguet62.test.springreactiveglobalcontext.threadlocal;
 
+import lombok.Getter;
+import lombok.Setter;
+
 public class ThreadLocalStorage {
 
+    @Getter
+    @Setter
     private static final ThreadLocal<String> threadLocal = new ThreadLocal();
 
     public static String getValue() {
@@ -11,5 +16,4 @@ public class ThreadLocalStorage {
     public static void setValue(String value) {
         threadLocal.set(value);
     }
-
 }
