@@ -2,6 +2,7 @@ package fr.pinguet62.test.springstatemachine.errorhandling;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.statemachine.StateContext;
 import org.springframework.statemachine.action.Action;
 
 @Configuration
@@ -9,9 +10,11 @@ public class TestActionsUtils {
 
     @Bean // proxied if necessary
     public Action<MyState, MyEvent> fallingAction() {
-        return c -> {
-            throw new RuntimeException("test error");
+        return new Action<MyState, MyEvent>() {
+            @Override
+            public void execute(StateContext<MyState, MyEvent> c) {
+                throw new RuntimeException("test error");
+            }
         };
     }
-
 }
