@@ -3,4 +3,5 @@ package fr.pinguet62.test.springstatemachine.errorhandling;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TestConfig {}
+public class TestConfig {
+}
