@@ -1,11 +1,10 @@
 package fr.pinguet62.test.springstatemachine.errorhandling.custom;
 
-import org.springframework.statemachine.StateMachine;
-import org.springframework.stereotype.Service;
-
 import fr.pinguet62.test.springstatemachine.errorhandling.ErrorInterceptorAspect.InterceptInternalStateMachineErrors;
 import fr.pinguet62.test.springstatemachine.errorhandling.MyEvent;
 import fr.pinguet62.test.springstatemachine.errorhandling.MyState;
+import org.springframework.statemachine.StateMachine;
+import org.springframework.stereotype.Service;
 
 @Service
 public class MyService {
@@ -14,5 +13,4 @@ public class MyService {
     public boolean sendSampleEvent(StateMachine<MyState, MyEvent> stateMachine) {
         return stateMachine.sendEvent(MyEvent.Sample);
     }
-
 }
