@@ -1,15 +1,16 @@
 package fr.pinguet62.test.utils.sax;
 
-import static java.nio.file.Files.newInputStream;
-import static java.nio.file.Paths.get;
+import org.w3c.dom.ls.LSInput;
+import org.w3c.dom.ls.LSResourceResolver;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.UncheckedIOException;
 import java.net.URI;
 import java.nio.file.Path;
 
-import org.w3c.dom.ls.LSInput;
-import org.w3c.dom.ls.LSResourceResolver;
+import static java.nio.file.Files.newInputStream;
+import static java.nio.file.Paths.get;
 
 /**
  * Implementation of {@link LSResourceResolver} who use only <b>relative path</b> to find included and imported other
@@ -24,9 +25,9 @@ public class LSResourceResolverImpl implements LSResourceResolver {
     }
 
     /**
-     * @param baseURI {@link URI} of XSD in process.
+     * @param baseURI  {@link URI} of XSD in process.
      * @param systemId Relative path (starting or not with {@code "./"} or {@code "../"}) of target included/imported
-     *        XSD. See {@code schemaLocation} tag attribute.
+     *                 XSD. See {@code schemaLocation} tag attribute.
      */
     @Override
     public LSInput resolveResource(String type, String namespaceURI, String publicId, String systemId, String baseURI) {
@@ -37,8 +38,7 @@ public class LSResourceResolverImpl implements LSResourceResolver {
         try {
             inputStream = newInputStream(systemIdPath);
         } catch (IOException e) {
-            // TODO Java 8: UncheckedIOException
-            throw new RuntimeException(e);
+            throw new UncheckedIOException(e);
         }
 
         LSInput lsInput = new LSInputImpl();
@@ -47,5 +47,4 @@ public class LSResourceResolverImpl implements LSResourceResolver {
         lsInput.setSystemId(systemIdPath.toUri().toString());
         return lsInput;
     }
-
 }
